Refine OAuth MCP domain actions
