+++ server/attachments/errorCodes.json
@@ -1,0 +1,24 @@
+{
+  "attachmentBatchLimitExceeded": "attachment_batch_limit_exceeded",
+  "attachmentMediaUnsupportedPrefix": "attachment_media_unsupported:",
+  "attachmentValidationFailed": "attachment_validation_failed",
+  "attachmentValidationErrorsPrefix": "attachment_validation_errors:",
+  "attachmentsRequired": "attachments_required",
+  "capabilityAttachmentUpload": "capability_required:attachment.upload",
+  "capabilityVideoUpload": "capability_required:attachment.video.upload",
+  "compressedFileInvalidPrefix": "compressed_file_invalid:",
+  "fileCountExceeded": "file_count_exceeded",
+  "insufficientVideoUpload": "insufficient_scope:video:upload",
+  "livePhotoOriginalNotImage": "live_photo_original_not_image",
+  "livePhotoPairedVideoInvalidPrefix": "live_photo_paired_video_invalid:",
+  "livePhotoPairedVideoMissingPrefix": "live_photo_paired_video_missing:",
+  "livePhotoPairedVideoNotVideo": "live_photo_paired_video_not_video",
+  "livePhotoPairedVideoRequired": "live_photo_paired_video_required",
+  "objectKeyInvalid": "object_key_invalid",
+  "originalFileNotFoundPrefix": "original_file_not_found:",
+  "pairedVideoLivePhotoOnlyPrefix": "paired_video_live_photo_only:",
+  "posterFileInvalidPrefix": "poster_file_invalid:",
+  "storageNotConfigured": "storage_not_configured",
+  "uuidMismatchPrefix": "attachment_uuid_mismatch:",
+  "videoCompressedKeyForbiddenPrefix": "video_compressed_key_forbidden:"
+}
