+++ server/mcp/errorCodes.json
@@ -4,49 +4,28 @@
   "articleNotFoundOrDenied": "article_not_found_or_denied",
   "attachmentBatchLimitExceeded": "attachment_batch_limit_exceeded",
   "attachmentIdsRequired": "attachment_ids_required",
-  "attachmentMediaUnsupportedPrefix": "attachment_media_unsupported:",
-  "attachmentValidationFailed": "attachment_validation_failed",
-  "attachmentValidationErrorsPrefix": "attachment_validation_errors:",
-  "attachmentsRequired": "attachments_required",
   "archivedBooleanRequired": "archived_boolean_required",
   "authRequired": "mcp_auth_required",
   "batchLimitExceeded": "batch_limit_exceeded",
-  "capabilityAttachmentUpload": "capability_required:attachment.upload",
-  "capabilityVideoUpload": "capability_required:attachment.video.upload",
-  "compressedFileInvalidPrefix": "compressed_file_invalid:",
   "dateRangeRequired": "date_range_required",
-  "fileCountExceeded": "file_count_exceeded",
   "idsRequired": "ids_required",
   "insufficientScope": "insufficient_scope",
-  "insufficientVideoUpload": "insufficient_scope:video:upload",
   "invalidCalendarDatePrefix": "invalid_calendar_date:",
   "invalidDateFormat": "invalid_date_format",
   "invalidOrigin": "invalid_origin",
   "invalidRequest": "invalid_request",
   "invalidUuidPrefix": "invalid_uuid:",
-  "livePhotoOriginalNotImage": "live_photo_original_not_image",
-  "livePhotoPairedVideoInvalidPrefix": "live_photo_paired_video_invalid:",
-  "livePhotoPairedVideoMissingPrefix": "live_photo_paired_video_missing:",
-  "livePhotoPairedVideoNotVideo": "live_photo_paired_video_not_video",
-  "livePhotoPairedVideoRequired": "live_photo_paired_video_required",
   "mcpToolFailed": "mcp_tool_failed",
   "methodNotFoundPrefix": "method_not_found:",
   "nonNegativeIntegerRequiredPrefix": "non_negative_integer_required:",
   "noteNotFound": "note_not_found",
   "noteOwnershipRequired": "note_ownership_required",
   "notePrivate": "note_private",
-  "objectKeyInvalid": "object_key_invalid",
-  "originalFileNotFoundPrefix": "original_file_not_found:",
-  "pairedVideoLivePhotoOnlyPrefix": "paired_video_live_photo_only:",
   "positiveLimitRequired": "positive_limit_required",
-  "posterFileInvalidPrefix": "poster_file_invalid:",
   "roteNotFound": "rote_not_found",
-  "storageNotConfigured": "storage_not_configured",
   "tagCountExceeded": "tag_count_exceeded",
   "tagLengthExceeded": "tag_length_exceeded",
   "toolDefinitionMissingPrefix": "mcp_tool_definition_missing:",
   "toolNameRequired": "tool_name_required",
-  "unknownToolPrefix": "unknown_tool:",
-  "uuidMismatchPrefix": "attachment_uuid_mismatch:",
-  "videoCompressedKeyForbiddenPrefix": "video_compressed_key_forbidden:"
+  "unknownToolPrefix": "unknown_tool:"
 }
